LayoutTests/html5lib/generated/run-tests7-write-expected.txt

CONSOLE MESSAGE: SyntaxError: Unexpected token '<'
../resources/tests7.dat: PASS
